mcp: a runtime skip cannot save a compile error
The cross platform lint job refused paths_test.go on Windows because
syscall.Mkfifo does not exist there. The test guarded itself with a
runtime.GOOS check, which reads as though Windows had been handled and
cannot work: the skip runs at run time, the missing symbol is a compile
error, and the package failed to typecheck before any test could decide
to skip itself.

The fifo test moves behind //go:build !windows, which is the guard that
acts early enough, and is how this package already separates open_unix.go
from open_windows.go. The now unreachable runtime skip goes with it rather
than being kept alongside a build tag that already excludes the platform.

Verified both directions: go vet exit 0 on unix and on GOOS=windows, and
the fifo test still RUNS rather than skipping on unix.
